# book-network-ui/src/openapi/openapi.json
{
     "openapi": "3.0.1",
     "info": {
          "title": "OpenApi specification - Book networking",
          "description": "OpenApi documentation for Spring Security example",
          "termsOfService": "Term of service",
          "contact": {
               "name": "dicka",
               "url": "sidicka.blogspot.co.id",
               "email": "dickanirwansyah@gmail.com"
          },
          "license": {
               "name": "License name",
               "url": "sidicka.blogspot.co.id"
          },
          "version": "1.0"
     },
     "servers": [
          {
               "url": "http://localhost:8081/api/v1",
               "description": "Local ENV"
          },
          {
               "url": "https://dickanirwansyah.aws.co.id/api/v1",
               "description": "Prod ENV"
          }
     ],
     "security": [
          {
               "bearerAuth": []
          }
     ],
     "paths": {
...
          "/auth/register": {
               "post": {
                    "tags": [
                         "Authentication"
                    ],
                    "operationId": "register",
                    "requestBody": {
                         "content": {
                              "application/json": {
                                   "schema": {
                                        "$ref": "#/components/schemas/RegistrationRequest"
                                   }
                              }
                         },
                         "required": true
                    },
                    "responses": {
                         "202": {
                              "description": "Accepted",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "object"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
          "/auth/authenticate": {
               "post": {
                    "tags": [
                         "Authentication"
                    ],
                    "operationId": "authenticate",
                    "requestBody": {
                         "content": {
                              "application/json": {
                                   "schema": {
                                        "$ref": "#/components/schemas/AuthenticationRequest"
                                   }
                              }
                         },
                         "required": true
                    },
                    "responses": {
                         "200": {
                              "description": "OK",
                              "content": {
                                   "application/json": {
                                        "schema": {
                                             "type": "object"
                                        }
                                   }
                              }
                         }
                    }
               }
          },
...
          "/auth/activate-account": {
               "get": {
                    "tags": [
                         "Authentication"
                    ],
                    "operationId": "activeAccount",
                    "parameters": [
                         {
                              "name": "token",
                              "in": "query",
                              "required": true,
                              "schema": {
                                   "type": "string"
                              }
                         }
                    ],
                    "responses": {
                         "200": {
                              "description": "OK"
                         }
                    }
               }
          }
     },
     "components": {
          "schemas": {
...
               "RegistrationRequest": {
                    "required": [
                         "email",
                         "firstname",
                         "lastname",
                         "password"
                    ],
                    "type": "object",
                    "properties": {
                         "firstname": {
                              "type": "string"
                         },
                         "lastname": {
                              "type": "string"
                         },
                         "email": {
                              "type": "string"
                         },
                         "password": {
                              "maxLength": 2147483647,
                              "minLength": 8,
                              "type": "string"
                         }
                    }
               },
               "AuthenticationRequest": {
                    "required": [
                         "email",
                         "password"
                    ],
                    "type": "object",
                    "properties": {
                         "email": {
                              "type": "string"
                         },
                         "password": {
                              "maxLength": 2147483647,
                              "minLength": 8,
                              "type": "string"
                         }
                    }
               },
...
          },
          "securitySchemes": {
               "bearerAuth": {
                    "type": "http",
                    "description": "JWT auth description",
                    "in": "header",
                    "scheme": "bearer",
                    "bearerFormat": "JWT"
               }
          }
     }
}